Remove spring-security remnants from seuranta-service
Authorization does not work there anyway...
Let's see if it would work with the new spring-security like this.
Authorization needs to be fixed separately, for example by porting
to CXF (either waiting for SSE support to land to CXF

  https://issues.apache.org/jira/browse/CXF-5855

or by porting the async-polling to work with some other method,
such as websockets).